server/requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-Migrate==4.0.5
Flask-CORS==4.0.0
Flask-JWT-Extended==4.5.3
psycopg2-binary==2.9.7
marshmallow==3.20.1
python-dotenv==1.0.0
Werkzeug==2.3.7
pytz==2024.1
gunicorn==21.2.0 